# openapi.json
{
  "components": {
    "schemas": {
...
      "UnauthorizedResponse": {
        "properties": {
          "detail": {
            "additionalProperties": true,
            "description": "FastAPI error envelope. For this API, typically `{ \"detail\": {\"error\":\"unauthorized\"} }`.",
            "examples": [
              {
                "error": "unauthorized"
              }
            ],
            "title": "Detail",
            "type": "object"
          }
        },
        "required": [
          "detail"
        ],
        "title": "UnauthorizedResponse",
        "type": "object"
      },
      "V2ActionError": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "V2ActionError",
        "type": "object"
      },
...
      "V2ErrorEnvelope": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          },
          "error": {
            "$ref": "#/components/schemas/V2ActionError"
          },
          "ok": {
            "const": false,
            "default": false,
            "title": "Ok",
            "type": "boolean"
          },
          "requestId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requestid"
          }
        },
        "required": [
          "error"
        ],
        "title": "V2ErrorEnvelope",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
...
    "title": "Hue Gateway",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/events/stream": {
      "get": {
        "description": "Server-Sent Events (SSE) stream of normalized Hue change events.\n\nResponse is `text/event-stream` where each event is sent as a single `data: <json>` frame.\nThe gateway may also send `: keepalive` comment frames.\n\nClients should:\n- keep the HTTP connection open\n- reconnect on disconnect\n- treat event `data` as best-effort (shape may evolve)",
        "operationId": "events_stream_v1_events_stream_get",
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "SSE stream (text/event-stream)."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponse"
                }
              }
            },
            "description": "Unauthorized."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          },
          {
            "APIKeyHeader": []
          }
        ],
        "summary": "Normalized Hue event stream (SSE)",
        "tags": [
          "events"
        ]
      }
    },
...
    "/v2/events/stream": {
      "get": {
        "operationId": "v2_events_stream_v2_events_stream_get",
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/V2ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "summary": "Event stream (v2, stub)",
        "tags": [
          "v2"
        ]
      }
    }
  }
}